refactor - move dns feature implementation to its own package
