+++ internal/clusterfeature/secrets_service.go
@@ -16,10 +16,6 @@ package clusterfeature
 
 import (
 	"context"
-
-	"emperror.dev/errors"
-	"github.com/banzaicloud/pipeline/secret"
-	"github.com/goph/logur"
 )
 
 // SecretsService component interface for secret handling abstraction
@@ -28,28 +24,4 @@ type SecretsService interface {
 	GetSecretValues(ctx context.Context, secretName string, orgID uint) (interface{}, error)
 }
 
-type secretsService struct {
-	logger logur.Logger
-}
 
-func (s *secretsService) GetSecretValues(ctx context.Context, secretName string, orgID uint) (interface{}, error) {
-
-	s.logger.Info("resolving secret ...", map[string]interface{}{"name": secretName, "orgID": orgID})
-	secret, err := secret.Store.GetByName(orgID, secretName)
-	if err != nil {
-		s.logger.Debug("failed to get secret", map[string]interface{}{"name": secretName, "orgID": orgID})
-
-		return nil, errors.WrapIf(err, "failed to get secret")
-	}
-
-	s.logger.Info("secret resolved", map[string]interface{}{"name": secretName, "orgID": orgID})
-	return secret.Values, nil
-
-}
-
-func NewSecretsService(logger logur.Logger) SecretsService {
-
-	return &secretsService{
-		logger: logur.WithFields(logger, map[string]interface{}{"secrets-service": "comp"}),
-	}
-}
